# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Vibe Cooking API
  description: Vibe CookingのためのAPI
  version: 1.0.0
  contact:
    name: ギャFUN!!
    email: gyafun@furari.co
servers:
  - url: http://localhost:3000/api
    description: Next.js
paths:
...
  /vibe-recipe:
    post:
      summary: バイブレシピ作成
      description: 指定されたレシピIDsからバイブレシピを作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeIds:
                  type: array
                  items:
                    type: string
                  description: バイブレシピに含めるレシピIDの配列
                  example: ['cm123abc456', 'cm789def012']
              required:
                - recipeIds
      responses:
        '200':
          description: 既存のバイブレシピを返します
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VibeRecipe'
        '201':
          description: バイブレシピが正常に作成されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VibeRecipe'
        '400':
          description: 無効なリクエスト
components:
  schemas:
...
    VibeRecipe:
      type: object
      properties:
        id:
          type: string
          description: VibeレシピID
        recipeIds:
          type: array
          items:
            type: string
          description: 関連するレシピIDの配列
        vibeInstructions:
          type: array
          items:
            $ref: '#/components/schemas/VibeInstruction'
          description: Vibe手順の配列
      required:
        - id
        - recipeIds
        - vibeInstructions
    VibeInstruction:
      type: object
      properties:
        instructionId:
          type: string
          description: 手順ID
        step:
          type: integer
          description: 手順番号
      required:
        - instructionId
        - step